Issue #11 Es treten keine Exception mehrs auf, wenn der Server keine Daten runtergibt. Es werden in diesem Fall auch keine Printjobs angelegt, welche unnötig wären. In diesen Fällen wird der Benutzer mit einem PopupDialog informiert.

+++ parent/client/src/main/java/ch/infbr5/sentinel/client/util/DesktopHelper.java
@@ -5,33 +5,41 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 
+import javax.swing.JOptionPane;
+
 public class DesktopHelper {
 
 	public static void openPdfFile(String filename, byte[] data) {
 
-		try {
-
-			// Create temp file.
-			File temp = File.createTempFile("sentinel_", ".pdf");
-
-			// Delete temp file when program exits.
-			temp.deleteOnExit();
-
-			// Write to temp file
-
-			FileOutputStream fos = new FileOutputStream(temp);
-			fos.write(data);
-			fos.close();
-
-			if (Desktop.isDesktopSupported()) {
-				Desktop.getDesktop().open(temp);
-			} else {
-				// Open Document Viewer in Lubuntu
-				Runtime.getRuntime().exec("evince " + temp.getAbsolutePath());
+		if (data == null) {
+			JOptionPane.showMessageDialog(null,
+					"Datei nicht vorhanden (leer).",
+					"Datei kann nicht geöffnet werden.", JOptionPane.WARNING_MESSAGE);
+		} else {
+			try {
+	
+				// Create temp file.
+				File temp = File.createTempFile("sentinel_", ".pdf");
+	
+				// Delete temp file when program exits.
+				temp.deleteOnExit();
+	
+				// Write to temp file
+	
+				FileOutputStream fos = new FileOutputStream(temp);
+				fos.write(data);
+				fos.close();
+	
+				if (Desktop.isDesktopSupported()) {
+					Desktop.getDesktop().open(temp);
+				} else {
+					// Open Document Viewer in Lubuntu
+					Runtime.getRuntime().exec("evince " + temp.getAbsolutePath());
+				}
+	
+			} catch (IOException e) {
+				e.printStackTrace();
 			}
-
-		} catch (IOException e) {
-			e.printStackTrace();
 		}
 
 	}
